Update captured Janrain attributes upon login
When users log in, their attributes from Janrain are persisted locally via the
set_capturable_params method. If one of these attributes is changed between
logins, the updated value is not persisted in the application's database upon
the next login. If the application using Devise::Capturable relies on any of
these values, it may be working with stale data as a result.

Fix by making sure set_capturable_params is called on a user that is already
persisted, before reporting a successful login within the Devise.

This behavior is not present in the current main repository for this Devise
plugin. In addition, we had to modify tests to keep existing functionality
passing. It's possible that the existing tests weren't completely exercising
the code. We intend to pursue this further and submit a pull request in the
future.