Make server-side redirect_uri value configurable
In order for Janrain to work, the value of redirect_uri set on the client and
server sides must match. Further, for the Federate option to work, this value
must match the protocol and domain of the application's environment. The
current implementation hardcodes this value, so it can't be customized per
application and environment.

Fix by making the value configurable per application and environment.

Task: https://oreillymedia.atlassian.net/browse/PT-560