Make Javascript overridable
The values for a few Janrain and authentication-related settings are hard-set in
the Javascript provided in the gem. These need to be customizable in order for
the following scenarios to work:

* To use the `signInEmbedded` flow instead of the default `signIn`
* To specify a `redirectUri` value (required to use Janrain's Federate feature)
* To add functionality to Janrain event handlers, or add more event handlers
* To use a different Devise resource than User

To fix, allow the Javascript to be overridden. A developer can replace the default `janrainCaptureWidgetOnLoad` function with one specific to the application. In addition, an optional path may be passed to the `afterJanrainLogin` function to specify a Devise resource to use when completing
login.